Added clone functions to Graphic2D and derivatives

+++ include/NeoInfused/Layer/Neo_Layer.hpp
@@ -7,7 +7,7 @@ namespace neo {
     class Layer {
         friend class LayerGroup;
     public:
-        Layer(float priority, bool enabled = true) : m_Priority(priority), m_Enabled(true) {}
+        Layer(int64_t priority, bool enabled = true) : m_Priority(priority), m_Enabled(true) {}
         virtual ~Layer(void) = default;
     public:
         virtual void update(void) {}
@@ -19,7 +19,7 @@ namespace neo {
         inline void disable(void) { m_Enabled = false; }
         inline void set_enabled(bool value) { m_Enabled = value; }
         inline bool enabled(void) const { return m_Enabled; }
-        inline float priority(void) const { return m_Priority; }
+        inline int64_t priority(void) const { return m_Priority; }
 
         inline bool operator>(const Layer& r)  const { return (m_Priority > r.m_Priority); }
         inline bool operator>=(const Layer& r) const { return (m_Priority >= r.m_Priority); }
@@ -28,12 +28,12 @@ namespace neo {
         inline bool operator==(const Layer& r) const { return (m_Priority == r.m_Priority); }
         inline bool operator!=(const Layer& r) const { return (m_Priority != r.m_Priority); }
 
-        inline operator float() const { return m_Priority; }
+        inline operator int64_t() const { return m_Priority; }
         inline operator bool() const { return m_Enabled; }
     protected:
         bool m_Enabled;
     private:
-        float m_Priority;
+        int64_t m_Priority;
     };
 } // namespace neo
